requirements.txt
flask
Flask-SQLAlchemy
flask-migrate
pymysql
flask-cors
requests
mysql-connector
flask-restplus
Flask-SQLAlchemy
pandas
google-cloud-storage
